--- server/rootvalue/rootvalue.go
@@ -1,723 +1,0 @@
-// Copyright 2024 Dolthub, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package rootvalue
-
-import (
-	"bytes"
-	"context"
-	"fmt"
-	"sort"
-	"strings"
-
-	"github.com/dolthub/dolt/go/libraries/doltcore/doltdb"
-	"github.com/dolthub/dolt/go/libraries/doltcore/doltdb/durable"
-	"github.com/dolthub/dolt/go/libraries/doltcore/schema"
-	"github.com/dolthub/dolt/go/store/hash"
-	"github.com/dolthub/dolt/go/store/prolly/tree"
-	"github.com/dolthub/dolt/go/store/types"
-)
-
-const (
-	ddbRootStructName = "dolt_db_root"
-	tablesKey         = "tables"
-	foreignKeyKey     = "foreign_key"
-	featureVersKey    = "feature_ver"
-	rootCollationKey  = "root_collation_key"
-)
-
-// DoltgresFeatureVersion is Doltgres' feature version. We use Dolt's feature version added to our own.
-var DoltgresFeatureVersion = doltdb.DoltFeatureVersion + 0
-
-// rootValue is Dolt's implementation of RootValue.
-type rootValue struct {
-	vrw  types.ValueReadWriter
-	ns   tree.NodeStore
-	st   fbRvStorage
-	fkc  *doltdb.ForeignKeyCollection // cache the first load
-	hash hash.Hash                    // cache first load
-}
-
-type tableEdit struct {
-	name doltdb.TableName
-	ref  *types.Ref
-
-	// Used for rename.
-	old_name string
-}
-
-func (root *rootValue) CreateDatabaseSchema(ctx context.Context, dbSchema schema.DatabaseSchema) (doltdb.RootValue, error) {
-	existingSchemas, err := root.st.GetSchemas(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, s := range existingSchemas {
-		if strings.EqualFold(s.Name, dbSchema.Name) {
-			return nil, fmt.Errorf("A schema with the name %s already exists", dbSchema.Name)
-		}
-	}
-
-	existingSchemas = append(existingSchemas, dbSchema)
-	sort.Slice(existingSchemas, func(i, j int) bool {
-		return existingSchemas[i].Name < existingSchemas[j].Name
-	})
-
-	r, err := root.st.SetSchemas(ctx, existingSchemas)
-	if err != nil {
-		return nil, err
-	}
-
-	return root.withStorage(r), nil
-}
-
-// CreateEmptyTable creates an empty table in this root with the name and schema given, returning the new root value.
-func (root *rootValue) CreateEmptyTable(ctx context.Context, tName doltdb.TableName, sch schema.Schema) (doltdb.RootValue, error) {
-	tbl, err := doltdb.CreateEmptyTable(ctx, root.NodeStore(), root.VRW(), sch)
-	if err != nil {
-		return nil, err
-	}
-
-	newRoot, err := root.PutTable(ctx, tName, tbl)
-	if err != nil {
-		return nil, err
-	}
-
-	return newRoot, nil
-}
-
-// DebugString returns a human readable string with the contents of this root. If |transitive| is true, row data from
-// all tables is also included. This method is very expensive for large root values, so |transitive| should only be used
-// when debugging tests.
-func (root *rootValue) DebugString(ctx context.Context, transitive bool) string {
-	var buf bytes.Buffer
-	buf.WriteString(root.st.DebugString(ctx))
-
-	if transitive {
-		buf.WriteString("\nTables:")
-		root.IterTables(ctx, func(name string, table *doltdb.Table, sch schema.Schema) (stop bool, err error) {
-			buf.WriteString("\nTable ")
-			buf.WriteString(name)
-			buf.WriteString(":\n")
-
-			buf.WriteString(table.DebugString(ctx, root.ns))
-
-			return false, nil
-		})
-	}
-
-	return buf.String()
-}
-
-func (root *rootValue) GenerateTagsForNewColColl(ctx context.Context, tableName string, cc *schema.ColCollection) (*schema.ColCollection, error) {
-	newColNames := make([]string, 0, cc.Size())
-	newColKinds := make([]types.NomsKind, 0, cc.Size())
-	_ = cc.Iter(func(tag uint64, col schema.Column) (stop bool, err error) {
-		newColNames = append(newColNames, col.Name)
-		newColKinds = append(newColKinds, col.Kind)
-		return false, nil
-	})
-
-	newTags, err := root.GenerateTagsForNewColumns(ctx, tableName, newColNames, newColKinds, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	idx := 0
-	return schema.MapColCollection(cc, func(col schema.Column) schema.Column {
-		col.Tag = newTags[idx]
-		idx++
-		return col
-	}), nil
-}
-
-// GenerateTagsForNewColumns deterministically generates a slice of new tags that are unique within the history of this root. The names and NomsKinds of
-// the new columns are used to see the tag generator.
-func (root *rootValue) GenerateTagsForNewColumns(
-	ctx context.Context,
-	tableName string,
-	newColNames []string,
-	newColKinds []types.NomsKind,
-	headRoot doltdb.RootValue,
-) ([]uint64, error) {
-	if len(newColNames) != len(newColKinds) {
-		return nil, fmt.Errorf("error generating tags, newColNames and newColKinds must be of equal length")
-	}
-
-	newTags := make([]*uint64, len(newColNames))
-
-	// Get existing columns from the current root, or the head root if the table doesn't exist in the current root. The
-	// latter case is to support reusing table tags in the case of drop / create in the same session, which is common
-	// during import.
-	existingCols, err := doltdb.GetExistingColumns(ctx, root, headRoot, tableName, newColNames, newColKinds)
-	if err != nil {
-		return nil, err
-	}
-
-	// If we found any existing columns set them in the newTags list.
-	for _, col := range existingCols {
-		col := col
-		for i := range newColNames {
-			// Only re-use tags if the noms kind didn't change
-			// TODO: revisit this when new storage format is further along
-			if strings.EqualFold(newColNames[i], col.Name) &&
-				newColKinds[i] == col.TypeInfo.NomsKind() {
-				newTags[i] = &col.Tag
-				break
-			}
-		}
-	}
-
-	var existingColKinds []types.NomsKind
-	for _, col := range existingCols {
-		existingColKinds = append(existingColKinds, col.Kind)
-	}
-
-	existingTags, err := doltdb.GetAllTagsForRoots(ctx, headRoot, root)
-	if err != nil {
-		return nil, err
-	}
-
-	outputTags := make([]uint64, len(newTags))
-	for i := range newTags {
-		if newTags[i] != nil {
-			outputTags[i] = *newTags[i]
-			continue
-		}
-
-		outputTags[i] = schema.AutoGenerateTag(existingTags, tableName, existingColKinds, newColNames[i], newColKinds[i])
-		existingColKinds = append(existingColKinds, newColKinds[i])
-		existingTags.Add(outputTags[i], tableName)
-	}
-
-	return outputTags, nil
-}
-
-func (root *rootValue) GetAllSchemas(ctx context.Context) (map[string]schema.Schema, error) {
-	m := make(map[string]schema.Schema)
-	err := root.IterTables(ctx, func(name string, table *doltdb.Table, sch schema.Schema) (stop bool, err error) {
-		m[name] = sch
-		return false, nil
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	return m, nil
-}
-
-func (root *rootValue) GetCollation(ctx context.Context) (schema.Collation, error) {
-	return root.st.GetCollation(ctx)
-}
-
-func (root *rootValue) GetDatabaseSchemas(ctx context.Context) ([]schema.DatabaseSchema, error) {
-	existingSchemas, err := root.st.GetSchemas(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	return existingSchemas, nil
-}
-
-// GetFeatureVersion returns the feature version of this root, if one is written
-func (root *rootValue) GetFeatureVersion(ctx context.Context) (ver doltdb.FeatureVersion, ok bool, err error) {
-	return root.st.GetFeatureVersion()
-}
-
-// GetForeignKeyCollection returns the ForeignKeyCollection for this root. As collections are meant to be modified
-// in-place, each returned collection may freely be altered without affecting future returned collections from this root.
-func (root *rootValue) GetForeignKeyCollection(ctx context.Context) (*doltdb.ForeignKeyCollection, error) {
-	if root.fkc == nil {
-		fkMap, ok, err := root.st.GetForeignKeys(ctx, root.vrw)
-		if err != nil {
-			return nil, err
-		}
-		if !ok {
-			return doltdb.NewForeignKeyCollection()
-		}
-
-		root.fkc, err = doltdb.DeserializeForeignKeys(ctx, root.vrw.Format(), fkMap)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return root.fkc.Copy(), nil
-}
-
-// GetTable will retrieve a table by its case-sensitive name.
-func (root *rootValue) GetTable(ctx context.Context, tName doltdb.TableName) (*doltdb.Table, bool, error) {
-	tableMap, err := root.getTableMap(ctx, tName.Schema)
-	if err != nil {
-		return nil, false, err
-	}
-
-	addr, err := tableMap.Get(ctx, tName.Name)
-	if err != nil {
-		return nil, false, err
-	}
-
-	return doltdb.GetTable(ctx, root, addr)
-}
-
-// GetTableByColTag looks for the table containing the given column tag.
-func (root *rootValue) GetTableByColTag(ctx context.Context, tag uint64) (tbl *doltdb.Table, name string, found bool, err error) {
-	err = root.IterTables(ctx, func(tn string, t *doltdb.Table, s schema.Schema) (bool, error) {
-		_, found = s.GetAllCols().GetByTag(tag)
-		if found {
-			name, tbl = tn, t
-		}
-
-		return found, nil
-	})
-
-	if err != nil {
-		return nil, "", false, err
-	}
-
-	return tbl, name, found, nil
-}
-
-func (root *rootValue) GetTableHash(ctx context.Context, tName string) (hash.Hash, bool, error) {
-	tableMap, err := root.getTableMap(ctx, doltdb.DefaultSchemaName)
-	if err != nil {
-		return hash.Hash{}, false, err
-	}
-
-	tVal, err := tableMap.Get(ctx, tName)
-	if err != nil {
-		return hash.Hash{}, false, err
-	}
-
-	return tVal, !tVal.IsEmpty(), nil
-}
-
-// GetTableInsensitive will retrieve a table by its case-insensitive name.
-func (root *rootValue) GetTableInsensitive(ctx context.Context, tName string) (*doltdb.Table, string, bool, error) {
-	resolvedName, ok, err := root.ResolveTableName(ctx, tName)
-	if err != nil {
-		return nil, "", false, err
-	}
-	if !ok {
-		return nil, "", false, nil
-	}
-	tbl, ok, err := root.GetTable(ctx, doltdb.TableName{Name: resolvedName})
-	if err != nil {
-		return nil, "", false, err
-	}
-	return tbl, resolvedName, ok, nil
-}
-
-// GetTableNames retrieves the lists of all tables for a RootValue
-func (root *rootValue) GetTableNames(ctx context.Context, schemaName string) ([]string, error) {
-	tableMap, err := root.getTableMap(ctx, schemaName)
-	if err != nil {
-		return nil, err
-	}
-
-	var names []string
-	err = tmIterAll(ctx, tableMap, func(name string, _ hash.Hash) {
-		names = append(names, name)
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return names, nil
-}
-
-func (root *rootValue) HasConflicts(ctx context.Context) (bool, error) {
-	cnfTbls, err := root.TablesWithDataConflicts(ctx)
-
-	if err != nil {
-		return false, err
-	}
-
-	return len(cnfTbls) > 0, nil
-}
-
-// HasConstraintViolations returns whether any tables have constraint violations.
-func (root *rootValue) HasConstraintViolations(ctx context.Context) (bool, error) {
-	tbls, err := root.TablesWithConstraintViolations(ctx)
-	if err != nil {
-		return false, err
-	}
-	return len(tbls) > 0, nil
-}
-
-// HashOf gets the hash of the root value
-func (root *rootValue) HashOf() (hash.Hash, error) {
-	if root.hash.IsEmpty() {
-		var err error
-		root.hash, err = root.st.nomsValue().Hash(root.vrw.Format())
-		if err != nil {
-			return hash.Hash{}, nil
-		}
-	}
-	return root.hash, nil
-}
-
-func (root *rootValue) HasTable(ctx context.Context, tName string) (bool, error) {
-	tableMap, err := root.st.GetTablesMap(ctx, root.vrw, root.ns, doltdb.DefaultSchemaName)
-	if err != nil {
-		return false, err
-	}
-	a, err := tableMap.Get(ctx, tName)
-	if err != nil {
-		return false, err
-	}
-	return !a.IsEmpty(), nil
-}
-
-// IterTables calls the callback function cb on each table in this RootValue.
-func (root *rootValue) IterTables(ctx context.Context, cb func(name string, table *doltdb.Table, sch schema.Schema) (stop bool, err error)) error {
-	// TODO: schema name
-	tm, err := root.getTableMap(ctx, doltdb.DefaultSchemaName)
-	if err != nil {
-		return err
-	}
-
-	return tm.Iter(ctx, func(name string, addr hash.Hash) (bool, error) {
-		nt, err := durable.TableFromAddr(ctx, root.VRW(), root.ns, addr)
-		if err != nil {
-			return true, err
-		}
-		tbl := doltdb.NewTableFromDurable(nt)
-
-		sch, err := tbl.GetSchema(ctx)
-		if err != nil {
-			return true, err
-		}
-
-		return cb(name, tbl, sch)
-	})
-}
-
-// MapTableHashes returns a map of each table name and hash.
-func (root *rootValue) MapTableHashes(ctx context.Context) (map[string]hash.Hash, error) {
-	// TODO: schema name
-	names, err := root.GetTableNames(ctx, doltdb.DefaultSchemaName)
-	if err != nil {
-		return nil, err
-	}
-	nameToHash := make(map[string]hash.Hash)
-	for _, name := range names {
-		h, ok, err := root.GetTableHash(ctx, name)
-		if err != nil {
-			return nil, err
-		} else if !ok {
-			return nil, fmt.Errorf("root found a table with name '%s' but no hash", name)
-		} else {
-			nameToHash[name] = h
-		}
-	}
-	return nameToHash, nil
-}
-
-func (root *rootValue) NodeStore() tree.NodeStore {
-	return root.ns
-}
-
-func (root *rootValue) NomsValue() types.Value {
-	return root.st.nomsValue()
-}
-
-// PutForeignKeyCollection returns a new root with the given foreign key collection.
-func (root *rootValue) PutForeignKeyCollection(ctx context.Context, fkc *doltdb.ForeignKeyCollection) (doltdb.RootValue, error) {
-	value, err := doltdb.SerializeForeignKeys(ctx, root.vrw, fkc)
-	if err != nil {
-		return nil, err
-	}
-	newStorage, err := root.st.SetForeignKeyMap(ctx, root.vrw, value)
-	if err != nil {
-		return nil, err
-	}
-	return root.withStorage(newStorage), nil
-}
-
-// PutTable inserts a table by name into the map of tables. If a table already exists with that name it will be replaced
-func (root *rootValue) PutTable(ctx context.Context, tName doltdb.TableName, table *doltdb.Table) (doltdb.RootValue, error) {
-	err := doltdb.ValidateTagUniqueness(ctx, root, tName.Name, table)
-	if err != nil {
-		return nil, err
-	}
-
-	tableRef, err := doltdb.RefFromNomsTable(ctx, table)
-	if err != nil {
-		return nil, err
-	}
-
-	return root.putTable(ctx, tName, tableRef)
-}
-
-func (root *rootValue) RemoveTables(ctx context.Context, skipFKHandling bool, allowDroppingFKReferenced bool, tables ...string) (doltdb.RootValue, error) {
-	// TODO: schema name
-	tableMap, err := root.getTableMap(ctx, doltdb.DefaultSchemaName)
-	if err != nil {
-		return nil, err
-	}
-
-	edits := make([]tableEdit, len(tables))
-	for i, name := range tables {
-		a, err := tableMap.Get(ctx, name)
-		if err != nil {
-			return nil, err
-		}
-		if a.IsEmpty() {
-			return nil, fmt.Errorf("%w: '%s'", doltdb.ErrTableNotFound, name)
-		}
-		edits[i].name = doltdb.TableName{
-			Name: name,
-		}
-	}
-
-	newStorage, err := root.st.EditTablesMap(ctx, root.vrw, root.ns, edits)
-	if err != nil {
-		return nil, err
-	}
-
-	newRoot := root.withStorage(newStorage)
-	if skipFKHandling {
-		return newRoot, nil
-	}
-
-	fkc, err := newRoot.GetForeignKeyCollection(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	if allowDroppingFKReferenced {
-		err = fkc.RemoveAndUnresolveTables(ctx, root, tables...)
-	} else {
-		err = fkc.RemoveTables(ctx, tables...)
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	return newRoot.PutForeignKeyCollection(ctx, fkc)
-}
-
-// RenameTable renames a table by changing its string key in the RootValue's table map. In order to preserve
-// column tag information, use this method instead of a table drop + add.
-func (root *rootValue) RenameTable(ctx context.Context, oldName, newName string) (doltdb.RootValue, error) {
-	newStorage, err := root.st.EditTablesMap(ctx, root.vrw, root.ns, []tableEdit{{old_name: oldName, name: doltdb.TableName{Name: newName}}})
-	if err != nil {
-		return nil, err
-	}
-	return root.withStorage(newStorage), nil
-}
-
-func (root *rootValue) ResolveRootValue(ctx context.Context) (doltdb.RootValue, error) {
-	return root, nil
-}
-
-// ResolveTableName resolves a case-insensitive name to the exact name as stored in Dolt. Returns false if no matching
-// name was found.
-func (root *rootValue) ResolveTableName(ctx context.Context, tName string) (string, bool, error) {
-	tableMap, err := root.getTableMap(ctx, doltdb.DefaultSchemaName)
-	if err != nil {
-		return "", false, err
-	}
-
-	a, err := tableMap.Get(ctx, tName)
-	if err != nil {
-		return "", false, err
-	}
-	if !a.IsEmpty() {
-		return tName, true, nil
-	}
-
-	found := false
-	err = tmIterAll(ctx, tableMap, func(name string, addr hash.Hash) {
-		if !found && strings.EqualFold(tName, name) {
-			tName = name
-			found = true
-		}
-	})
-	if err != nil {
-		return "", false, nil
-	}
-	return tName, found, nil
-}
-
-func (root *rootValue) SetCollation(ctx context.Context, collation schema.Collation) (doltdb.RootValue, error) {
-	newStorage, err := root.st.SetCollation(ctx, collation)
-	if err != nil {
-		return nil, err
-	}
-	return root.withStorage(newStorage), nil
-}
-
-func (root *rootValue) SetFeatureVersion(v doltdb.FeatureVersion) (doltdb.RootValue, error) {
-	newStorage, err := root.st.SetFeatureVersion(v)
-	if err != nil {
-		return nil, err
-	}
-	return root.withStorage(newStorage), nil
-}
-
-func (root *rootValue) SetTableHash(ctx context.Context, tName string, h hash.Hash) (doltdb.RootValue, error) {
-	val, err := root.vrw.ReadValue(ctx, h)
-
-	if err != nil {
-		return nil, err
-	}
-
-	ref, err := types.NewRef(val, root.vrw.Format())
-
-	if err != nil {
-		return nil, err
-	}
-
-	// TODO: schema
-	return root.putTable(ctx, doltdb.TableName{Name: tName}, ref)
-}
-
-// TablesWithConstraintViolations returns all tables that have constraint violations.
-func (root *rootValue) TablesWithConstraintViolations(ctx context.Context) ([]string, error) {
-	// TODO: schema name
-	names, err := root.GetTableNames(ctx, doltdb.DefaultSchemaName)
-	if err != nil {
-		return nil, err
-	}
-
-	violating := make([]string, 0, len(names))
-	for _, name := range names {
-		tbl, _, err := root.GetTable(ctx, doltdb.TableName{Name: name})
-		if err != nil {
-			return nil, err
-		}
-
-		n, err := tbl.NumConstraintViolations(ctx)
-		if err != nil {
-			return nil, err
-		}
-
-		if n > 0 {
-			violating = append(violating, name)
-		}
-	}
-
-	return violating, nil
-}
-
-func (root *rootValue) TablesWithDataConflicts(ctx context.Context) ([]string, error) {
-	names, err := root.GetTableNames(ctx, doltdb.DefaultSchemaName)
-	if err != nil {
-		return nil, err
-	}
-
-	conflicted := make([]string, 0, len(names))
-	for _, name := range names {
-		tbl, _, err := root.GetTable(ctx, doltdb.TableName{Name: name})
-		if err != nil {
-			return nil, err
-		}
-
-		ok, err := tbl.HasConflicts(ctx)
-		if err != nil {
-			return nil, err
-		}
-		if ok {
-			conflicted = append(conflicted, name)
-		}
-	}
-
-	return conflicted, nil
-}
-
-// ValidateForeignKeysOnSchemas ensures that all foreign keys' tables are present, removing any foreign keys where the declared
-// table is missing, and returning an error if a key is in an invalid state or a referenced table is missing. Does not
-// check any tables' row data.
-func (root *rootValue) ValidateForeignKeysOnSchemas(ctx context.Context) (doltdb.RootValue, error) {
-	fkCollection, err := root.GetForeignKeyCollection(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	// TODO: schema name
-	allTablesSlice, err := root.GetTableNames(ctx, doltdb.DefaultSchemaName)
-	if err != nil {
-		return nil, err
-	}
-	allTablesSet := make(map[string]schema.Schema)
-	for _, tableName := range allTablesSlice {
-		tbl, ok, err := root.GetTable(ctx, doltdb.TableName{Name: tableName})
-		if err != nil {
-			return nil, err
-		}
-		if !ok {
-			return nil, fmt.Errorf("found table `%s` in staging but could not load for foreign key check", tableName)
-		}
-		tblSch, err := tbl.GetSchema(ctx)
-		if err != nil {
-			return nil, err
-		}
-		allTablesSet[tableName] = tblSch
-	}
-
-	// some of these checks are sanity checks and should never happen
-	allForeignKeys := fkCollection.AllKeys()
-	for _, foreignKey := range allForeignKeys {
-		tblSch, existsInRoot := allTablesSet[foreignKey.TableName]
-		if existsInRoot {
-			if err := foreignKey.ValidateTableSchema(tblSch); err != nil {
-				return nil, err
-			}
-			parentSch, existsInRoot := allTablesSet[foreignKey.ReferencedTableName]
-			if !existsInRoot {
-				return nil, fmt.Errorf("foreign key `%s` requires the referenced table `%s`", foreignKey.Name, foreignKey.ReferencedTableName)
-			}
-			if err := foreignKey.ValidateReferencedTableSchema(parentSch); err != nil {
-				return nil, err
-			}
-		} else {
-			if !fkCollection.RemoveKeyByName(foreignKey.Name) {
-				return nil, fmt.Errorf("`%s` does not exist as a foreign key", foreignKey.Name)
-			}
-		}
-	}
-
-	return root.PutForeignKeyCollection(ctx, fkCollection)
-}
-
-func (root *rootValue) VRW() types.ValueReadWriter {
-	return root.vrw
-}
-
-func (root *rootValue) getTableMap(ctx context.Context, schemaName string) (tableMap, error) {
-	if schemaName == "" {
-		schemaName = doltdb.DefaultSchemaName
-	}
-	return root.st.GetTablesMap(ctx, root.vrw, root.ns, schemaName)
-}
-
-func (root *rootValue) putTable(ctx context.Context, tName doltdb.TableName, ref types.Ref) (doltdb.RootValue, error) {
-	if !doltdb.IsValidTableName(tName.Name) {
-		panic("Don't attempt to put a table with a name that fails the IsValidTableName check")
-	}
-
-	newStorage, err := root.st.EditTablesMap(ctx, root.VRW(), root.NodeStore(), []tableEdit{{name: tName, ref: &ref}})
-	if err != nil {
-		return nil, err
-	}
-
-	return root.withStorage(newStorage), nil
-}
-
-func (root *rootValue) withStorage(st fbRvStorage) *rootValue {
-	return &rootValue{root.vrw, root.ns, st, nil, hash.Hash{}}
-}
